requirements.txt
langgraph 
langgraph-sdk 
langgraph-checkpoint-sqlite 
langsmith 
langchain-community 
langchain-core 
langchain-openai 
tavily-python 
wikipedia 
google-generativeai